requirements.txt
langchain
langchain_community
langchain-openai
beautifulsoup4
python-dotenv
faiss-cpu
fastapi
fastapi-cors
uvicorn

-e .